requirements.txt
selenium
webdriver-manager
langchain-groq
langchain-core
python-dotenv
pydantic
